form works but chart is in trouble
