+++ budgetdb/views/__init__.py
@@ -1,0 +1,3 @@
+from .budgetedevent_views import *
+from .transaction_views import *
+from .views import *
